Used context in Frontend and made small changes in controller in Backend
