+++ Frontend/car-rental/src/pages/Home.jsx
@@ -1,9 +1,14 @@
+import { useContext } from "react";
 import Featured from "../components/Feature";
 import Footer from "../components/Footer";
 import Hero from "../components/Hero";
 import Testimonials from "../components/Review";
+import { AuthContext } from "../context/authContext";
 
-const Home = ({ userType, isLoggedIn }) => {
+const Home = () => {
+
+  const {userType,isLoggedIn} = useContext(AuthContext);
+
   return (
     <>
       <Hero userType={userType} isLoggedIn={isLoggedIn}></Hero>
